+++ pint2vela/pint2vela/convert_components.py
@@ -53,11 +53,16 @@ def pint_components_to_vela(model: TimingModel, toas: TOAs):
     elif "AstrometryEquatorial" in component_names:
         components.append(vl.SolarSystem(False, model.PLANET_SHAPIRO.value))
 
-    if "SolarWindDispersion" in component_names and not (model.NE_SW.value == 0 and model.NE_SW.frozen):
+    if "SolarWindDispersion" in component_names and not (
+        model.NE_SW.value == 0 and model.NE_SW.frozen
+    ):
         components.append(vl.SolarWindDispersion())
 
     if "DispersionDM" in component_names:
         components.append(vl.DispersionTaylor())
+
+    if "ChromaticCM" in component_names:
+        components.append(vl.ChromaticTaylor())
 
     if "Spindown" in component_names:
         components.append(vl.Spindown())