+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - More methods for `get_free_param_names()`, `read_param_values_to_vector()`, and `get_scale_factors()` for convenience
 - Added CI tests for `pint2vela`
 - An alternative implementation of `PhaseJump` for mutually exclusive JUMPs
-- Simple solar wind model (Edwards+ 2006)
+- Simple solar wind model (Edwards+ 2006) (`SolarWind`)
+- Variable-index chromatic delay as a Taylor expansion (`ChromaticTaylor`)
 - Auto-generation of HTML documentation using `Documenter`
 ## Changed
 - Reorganized source files into subdirectories